requirements.txt
streamlit==1.40.0
anthropic==0.39.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==11.0.0
pandas==2.2.3
